+++ src/01-Build/NukeBuild/Directory.Build.targets
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+	<!-- This file prevents unintended imports of unrelated MSBuild files -->
+	<!-- Uncomment to include parent Directory.Build.targets file -->
+	<!--<Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.targets', '$(MSBuildThisFileDirectory)../'))" />-->
+</Project>
