+++ src/01-Build/NukeBuild/NukeBuild.csproj
@@ -5,8 +5,8 @@
     <TargetFramework>net8.0</TargetFramework>
     <RootNamespace></RootNamespace>
     <NoWarn>CS0649;CS0169;CA1050;CA1822;CA2211;IDE1006</NoWarn>
-    <NukeRootDirectory>..\..</NukeRootDirectory>
-    <NukeScriptDirectory>..\..</NukeScriptDirectory>
+    <NukeRootDirectory>..\..\..</NukeRootDirectory>
+    <NukeScriptDirectory>..\..\..</NukeScriptDirectory>
     <NukeTelemetryVersion>1</NukeTelemetryVersion>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
@@ -17,6 +17,10 @@
 	<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
     <PackageReference Include="Nuke.Common" Version="8.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageDownload Include="nbgv" Version="[3.6.133]" />
   </ItemGroup>
 
 </Project>
