+++ src/01-Build/NukeBuild/NukeBuild.csproj.DotSettings
