+++ src/Wtq.sln
@@ -11,8 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "20-Host", "20-Host", "{F414
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wtq.Core", "10-Core\Wtq.Core\Wtq.Core.csproj", "{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NukeBuild", "01-Build\NukeBuild\NukeBuild.csproj", "{07080ACA-6607-4D56-8F54-DA33AB49AA9F}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wtq.Windows", "20-Host\Wtq.Windows\Wtq.Windows.csproj", "{87DC503F-11E4-48B7-BAD0-5B93B323009B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wtq.SharpHook", "10-Core\Wtq.SharpHook\Wtq.SharpHook.csproj", "{48C6BD6F-561C-41FE-B171-7B7AA96F421D}"
@@ -35,7 +33,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "00-Sln", "00-Sln", "{79182C
 		..\stylecop.json = ..\stylecop.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NukeBuild", "..\01-Build\NukeBuild.csproj", "{24CF68A7-4150-4F4B-9716-5F87611582A8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NukeBuild", "01-Build\NukeBuild\NukeBuild.csproj", "{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,16 +41,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{24CF68A7-4150-4F4B-9716-5F87611582A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{24CF68A7-4150-4F4B-9716-5F87611582A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330}.Release|Any CPU.Build.0 = Release|Any CPU
-		{07080ACA-6607-4D56-8F54-DA33AB49AA9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{07080ACA-6607-4D56-8F54-DA33AB49AA9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{07080ACA-6607-4D56-8F54-DA33AB49AA9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{07080ACA-6607-4D56-8F54-DA33AB49AA9F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{87DC503F-11E4-48B7-BAD0-5B93B323009B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{87DC503F-11E4-48B7-BAD0-5B93B323009B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{87DC503F-11E4-48B7-BAD0-5B93B323009B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -73,18 +65,22 @@ Global
 		{631C8A73-5698-4280-AE9D-D607BEEF4632}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{631C8A73-5698-4280-AE9D-D607BEEF4632}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{631C8A73-5698-4280-AE9D-D607BEEF4632}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{DCFC9B65-12AD-43B1-864C-1CEC0FCCF330} = {7DA42A83-90F6-40FF-B60D-673A2D2DAEAC}
-		{07080ACA-6607-4D56-8F54-DA33AB49AA9F} = {128FD100-EDD5-48CA-BE9B-7E1CC419C19E}
 		{87DC503F-11E4-48B7-BAD0-5B93B323009B} = {F414F7B7-3276-46DE-B47E-9DA69D56736F}
 		{48C6BD6F-561C-41FE-B171-7B7AA96F421D} = {7DA42A83-90F6-40FF-B60D-673A2D2DAEAC}
 		{35A4F38A-C249-41C8-8CE1-574F66104D60} = {7DA42A83-90F6-40FF-B60D-673A2D2DAEAC}
 		{0595151F-4EE9-456A-B1D3-42795F4D4ACC} = {7DA42A83-90F6-40FF-B60D-673A2D2DAEAC}
 		{631C8A73-5698-4280-AE9D-D607BEEF4632} = {7DA42A83-90F6-40FF-B60D-673A2D2DAEAC}
+		{36A15BA9-3AEF-49D6-9AE0-11E9F045EC55} = {128FD100-EDD5-48CA-BE9B-7E1CC419C19E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {65349C7B-F6B3-49A8-9935-4ABF9B6D3D70}
